+ moved goodgame dialog to separate file
